move metrics logic to integration test package and simplify test assertions
